refactor: co-locate legacy pde models with engines
