+++ src/pricing/engines/__init__.py
@@ -1,7 +1,10 @@
 """Pricing engines package."""
 
 from .finite_difference import (
+    BlackScholesPDE,
+    CallableBondPDEModel,
     GridParameters,
+    PDEModel,
     PricingEngine,
     PricingResult,
     create_default_pricing_engine,
@@ -15,6 +18,7 @@ from .unified import (
 
 __all__ = [
     "GridParameters",
+    "PDEModel",
     "PricingEngine",
     "PricingResult",
     "UnifiedPricingEngine",
@@ -22,4 +26,6 @@ __all__ = [
     "create_linear_grid",
     "create_log_grid",
     "create_unified_pricing_engine",
+    "BlackScholesPDE",
+    "CallableBondPDEModel",
 ]
