+++ tests/test_callable_bond.py
@@ -9,7 +9,7 @@ sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
 import numpy as np
 
 from src.pricing import OptionPricer
-from src.pde_pricer import CallableBondPDEModel
+from src.pricing.engines import CallableBondPDEModel
 from src.market import Market
 from src.processes import GeometricBrownianMotion
