+++ tests/test_pde_pricer.py
@@ -8,7 +8,7 @@ sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
 import numpy as np
 from src.processes.affine import GeometricBrownianMotion
 from src.instruments.base import EuropeanCall, EuropeanPut
-from src.pde_pricer import BlackScholesPDE
+from src.pricing.engines import BlackScholesPDE
 from src.time_steppers import ExplicitEuler
